+++ TheGuideToTheNewEden/TheGuideToTheNewEden.WinUI/TheGuideToTheNewEden.WinUI/Converters/UnreadMailVisibilityConverter.cs
@@ -12,7 +12,7 @@ namespace TheGuideToTheNewEden.WinUI.Converters
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            return (int)value > 0 ? Visibility.Visible : Visibility.Collapsed;
+            return (long)value > 0 ? Visibility.Visible : Visibility.Collapsed;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
